+++ src/plone/restapi/tests/test_services_users.py
@@ -687,6 +687,39 @@ class TestUsersEndpoint(unittest.TestCase):
 
         self.assertEqual(response.status_code, 200)
 
+    def test_reset_and_login_email_using_mail(self):
+        # enable use_email_as_login
+        security_settings = getAdapter(self.portal, ISecuritySchema)
+        security_settings.use_email_as_login = True
+        transaction.commit()
+
+        response = self.api_session.post(
+            '/@users',
+            json={
+                "email": "howard.zinn@example.com",
+                "password": "secret"
+            },
+        )
+        transaction.commit()
+
+        self.assertEqual(201, response.status_code)
+        user_id = response.json()['id']
+        user = api.user.get(userid=user_id)
+        self.assertTrue(user)
+
+        reset_tool = getToolByName(self.portal, 'portal_password_reset')
+        reset_info = reset_tool.requestReset(user.id)
+        token = reset_info['randomstring']
+        transaction.commit()
+
+        payload = {'reset_token': token,
+                   'new_password': 'new_password'}
+        response = self.api_session.post(
+            '/@users/{}/reset-password'.format(user.getUserName()),
+            json=payload)
+
+        self.assertEqual(response.status_code, 200)
+
     def test_delete_user(self):
         response = self.api_session.delete('/@users/noam')
         transaction.commit()